cleancode整改

Created-by: ninihihi
Commit-by: zouyanlong
Merged-by: turing_project1
Description: ## 描述
cleancode 8.5
## checklist
<!--  [x] 表示选中 -->
- [x] 是否通过本地IDE对代码进行静态检查
- [x] 是否通过本地IDE对代码进行格式化处理
- [x] 是否进行空指针校验
- [x] 是否进行返回值校验
- [x] 是否正确释放new/malloc申请的内存
- [x] 是否充分考虑接口的异常场景
- [x] 是否正确记录错误日志

See merge request: cann/ascend-transformer-boost!1780

+++ src/ops/ops_infer/layer_norm/layer_norm_aclnn_runner.cpp
@@ -97,10 +97,10 @@ aclnnStatus LayerNormAclnnRunner::SetAclNNWorkspaceExecutor()
     Dims &xDims = xAclnnTensorPtr->atbTensor.desc.shape;
     int32_t beginNormAxis = param_.normParam.beginNormAxis;
     if (beginNormAxis < 0) {
-        beginNormAxis += xDims.dimNum;
+        beginNormAxis += static_cast<int32_t>(xDims.dimNum);
     }
 
-    int64_t totalNormAxes = xDims.dimNum - beginNormAxis;
+    int64_t totalNormAxes = static_cast<int64_t>(xDims.dimNum) - beginNormAxis;
     int64_t shapeHostData[totalNormAxes];
     for (int i = 0; i < totalNormAxes; ++i) {
         shapeHostData[i] = xDims.dims[beginNormAxis + i];
